Move impl Codec for Certificate into closer to definition
